Prototype of a "relay" component
This change prototypes a relay component. The goal of this component
is to be a drop-in replacement for jaeger-collector, but instead of
writing to cassandra (or another writer) relay the spans to another
jaeger-collector (or compatible) instance.

Currently, the relay is its own component. In future, it could be
potentially merged into the jaeger-collector repo.