Do not use IncludeTypeName on ES7
Without IncludeTypeName, ES will use the type "_doc" all documents.

Type is removed from all search queries so that it queries all types
whether they be "_doc" or (eg) "span".

Type should not be used when indexing documents on ES7 (so that the
default "_doc" type is used).  It should, however, be used on ES5 and ES6
so that the type matches that described by the mapping.

Signed-off-by: Greg Franklin <gregoryfranklin@rentalcars.com>