WIP: Support for built-in status checking
This adds a new subcommand `status` which will simply query the admin
health endpoint and return success or failure.

Some container runtime environments (ECS) only support CMD healthchecks,
since Jaeger is built of off empty containers this effectively makes it
impossible for the orchestration engine to validate the container
started successfully. By adding a simple built-in status check we can
make it easy for a service (or operator) to validate the service is
running.

Signed-off-by: Daniel Kimsey <dkimsey@trustwave.com>