Fix file descriptor leaks in tlscfg unit tests
There are a few file descriptor leaks in the unit tests, and on an error
condition in certwatcher.

This is an issue when checking for flaky tests with
`go test -test.count ..`. One leak is in an error condition in
certwatcher but it is unlikely to be a runtime issue because.

Related #2586

Signed-off-by: Carl Henrik Lunde <chlunde@ifi.uio.no>