Add instrumentation handler to collector endpoints
Currently we do not expose any HTTP metrics on any of the collector
endpoints. This makes it harder to detect when the collector
is having issues ingesting bad spans for example. This change adds a
metric that reports the duration of each request and adds
HTTP status and path labels to it.

Signed-off-by: Dimitar Dimitrov <hi@dimitardimitrov.dev>