Add metrics instrumentation to collector endpoints (#2664)
* Add instrumentation handler to collector endpoints

Currently we do not expose any HTTP metrics on any of the collector
endpoints. This makes it harder to detect when the collector
is having issues ingesting bad spans for example. This change adds a
metric that reports the duration of each request and adds
HTTP status and path labels to it.

Signed-off-by: Dimitar Dimitrov <hi@dimitardimitrov.dev>

* Fix static check problems

Signed-off-by: Dimitar Dimitrov <hi@dimitardimitrov.dev>

* Add godoc for NewMetricsHandler

Signed-off-by: Dimitar Dimitrov <hi@dimitardimitrov.dev>

* Fix import formatting

Signed-off-by: Dimitar Dimitrov <hi@dimitardimitrov.dev>

* Up test coverage

Signed-off-by: Dimitar Dimitrov <hi@dimitardimitrov.dev>

* Rename packages and update license statements

Signed-off-by: Dimitar Dimitrov <hi@dimitardimitrov.dev>

* Up test coverage for server package

We weren't testing the starting of collector's HTTP and HTTP zipkin
servers. Added sanity tests for that. Also those two handlers
exited with log.Fatal, which made them hard to test; this has been
changed as well

Signed-off-by: Dimitar Dimitrov <hi@dimitardimitrov.dev>

* Cache metric timers in httpmetrics

Only the prometheus metrics.Factory implementaiton caches the
instantiated Timers, Gauges, etc. The other implementations
allocate new objects each time. We want to prevent that because
this is in the hot path. This adds caching on the request
duration timers. This shouldn't lead to noticeable increase in
memory usage.

Signed-off-by: Dimitar Dimitrov <hi@dimitardimitrov.dev>

* Add synchronisation in httpmetrics

It was possible that we get concurrent map read-write panics.

Signed-off-by: Dimitar Dimitrov <hi@dimitardimitrov.dev>

* Change server tests to use ephemeral ports

It's more robust if we use the next available port instead of
hardcoding a large port.

Signed-off-by: Dimitar Dimitrov <hi@dimitardimitrov.dev>

* Replace string with struct keys in httpmetrics

We can use directly struct keys instead of building strings out of each
struct we get.

Signed-off-by: Dimitar Dimitrov <hi@dimitardimitrov.dev>

* Increase await timeout in httmetrics test

Signed-off-by: Dimitar Dimitrov <hi@dimitardimitrov.dev>

* Simplify network calls in server tests

Signed-off-by: Dimitar Dimitrov <hi@dimitardimitrov.dev>

* Consolidate port strings in server tests

Signed-off-by: Dimitar Dimitrov <hi@dimitardimitrov.dev>

* Clean up a Stringer.String() and zero value init

Signed-off-by: Dimitar Dimitrov <hi@dimitardimitrov.dev>

* Use httptest in collector/app/server tests

Signed-off-by: Dimitar Dimitrov <hi@dimitardimitrov.dev>