Updated gRPC Storage Plugin README with example
I was struggling to generate the gRPC Storage Plugin bindings from the
storage.proto in the Jaeger repo so I created an question on the repo.
Yuri Shkuro helped me out by providing some additional suggestions and
information and requested that I add what I discovered to the README.
Therefore, this update is just a documentation update to the gRPC
Storage Plugin README.md to include details on how one can utilize the
Docker Protobuf image to generate gRPC bindings in C# (which is what my
use-case was).

Signed-off-by: Justin Stauffer <justin.stauffer@gmail.com>