Introduce proxy
Jaeger proxy is a new component which can use multiple jaeger query
instances as upstreams and merges the result of queries.
This component can be useful in island architecture setups where applications
communicate across islands but jaeger service is provided locally per island.

The implementation is based on the query component and also reuses the ui.
The communication between proxy and query upstreams is based on GRPC.
All queries are performed in parallel and results are merged and returned.
The proxy also performs adjustments on traces.

The proxy component expects the proper tags to be applied during collection.
This is important to identify the source of a span (e.g. `cluster=east`).

As an extra feature it provides an option to map existing tags to new ones:
```
source tag:
hostname: app-abcd1234

cli flag:
--proxy.tags.map hostname:pod

result:
hostname: app-abcd1234
pod: app-abcd1234
```
This is useful when integrating Jaeger datasource with other data sources
which might be using different labels (`pod` instead of `hostname`, etc.)
e.g. Grafana Loki.

This is an initial implementation as a proposal so some work still remains:
- Fixing applicable tests which were copied from query component
- Adding tests specific to proxy component
- Handle GRPC client options

Signed-off-by: Siavash Safi <siavash.safi@gmail.com>