Change build info date to commit timestamp (#3876)
This change fixes binary reproducibility of builds. Injecting a
timestamp of the time-of-build produces unique binaries for every build
which is undesirable for security and provenance tracking. If timestamps
for builds are desired, it is recommended to inject the timestamp of the
commit. This gives a timestamp that is consistent for a build based on
that commit and allows checking for binary consistenty across build systems.

This change updates BuildDate to use the commit timestamp from git.
This change also fixes some builds which include the version package but
were not injecting BUILD_INFO.

Signed-off-by: Calvin Behling <calvin.behling@gmail.com>

Signed-off-by: Calvin Behling <calvin.behling@gmail.com>