Change to shasum perl script
The perl package shasum is more likely to exist on development machines
than the GNU coreutils sha256sum and similar packages. Switching to the
shasum package should provide less development friction by not
introducing a new dependency.
Enable binary mode on the shasum script. GNU coreutils sha256sum uses
binary mode by default. It shouldn't make a significant difference to
users but I think it's best to just do it. This means updating the sed
script in the package-deploy script to ignore the leading asterisk on
file paths.
Added a guard to the package-deploy script around not detecting a
version name because I found that the script would silently accept an
empty string when the Makefile was invalid.

Signed-off-by: Calvin Behling <calvin.behling@gmail.com>