Use all span kinds as default for metrics
Instead of just using "SPAN_KIND_SERVER" as the default span kind for
the metrics, the default value now includes all span kinds.

Background: The span kind for some services could have another value
then "SPAN_KIND_SERVER", e.g. the spans from Istio are using
"SPAN_KIND_CLIENT". Because of the default span kind and the fact that
the span kind is not configureable in the frontend (the query parameter
is never set in the API calls) it was not possible to use the service
performance monitoring with some services.

Signed-off-by: ricoberger <mail@ricoberger.de>