Watch directories for certificate hot-reload
Instead of adding inotify watches on files, which mostly get removed during
certificate renewal, add watches for the parent directories and consider
inotify events only as indication that some of the files might have changed.

The change is required for the most typical atomic file update schemes,
for example for the scheme used by Kubernetes to update secret volumes.

Signed-off-by: Tero Saarni <tero.saarni@est.tech>