Watch directories for certificate hot-reload (#4159)
## Which problem is this PR solving?
This PR fixes the certificate hot-reload on Kubernetes.

Resolves #3968

## Short description of the changes
This PR moves the inotify watch from certificate files to monitoring the
parent directory instead, as also recommended
[here](https://github.com/fsnotify/fsnotify/blob/c6f5cfa163edb0f1bb78be3a77053ee14c48a3ce/backend_inotify.go#L246-L254)
by the [fsnotify](https://github.com/fsnotify/fsnotify) library.

Without this change, certificate hot-reload only works if the content of
certificate or key files is overwritten. That approach is rarely used
for update since it is not atomic and it risks files being read in the
middle of the update. Therefore more often the files are replaced by a
rename operation, by swapping the old file with a new one. It allows the
file to be completely written and closed before exposing it to the
application. However, in this update scheme, the old file gets deleted
and the current inotify watch on file level is automatically terminated.
No reload happens and warning message like following was printed into
log

```json
{"msg":"Certificate has been removed, using the last known version","certificate":"/certs/server-ca.pem"}
```

Example use case:

When Kubelet writes certificate and key files on Kubernetes Secret
volume, the target files are symbolic links to files in a different
directory - a hidden subdirectory in a same level as the files
themselves. During update, that directory is swapped with a new one,
while the symbolic links remains the same. This guarantees atomic swap
for both certificate and key files at once. It also means that any
rename event received at the parent directory level might indicate that
the files were replaced, even if name of the renamed file was not any of
the files being monitored. Therefore this PR proposes checking the
hashes of the files to detect changes.

With this change, following update schemes will work:
* Replace file content in-place (old behaviour: e.g. `cat newfile >
cert.pem`)
* Update the files by making any rename operation in the parent
directory (Kubernetes secret volume update scheme: `mv ..data_tmp
..data`)
* Update the files by renaming the target files (e.g. `mv cert.pem.tmp
cert.pem`)

Signed-off-by: Tero Saarni <tero.saarni@est.tech>

Signed-off-by: Tero Saarni <tero.saarni@est.tech>
Signed-off-by: Yuri Shkuro <github@ysh.us>
Co-authored-by: Yuri Shkuro <github@ysh.us>