Change go.mod's `go` directive to indicate support for Go 1.19
This partially reverts #4206.

On the OpenTelemetry Collector project we aim to provide support
for all officially supported Go versions. Per Go's support policy [1],
Go 1.19 will remain a supported Go version until Go 1.21 is released.
Since we take a dependency on Jaeger as a library, we need Jaeger to
remain compatible with Go 1.19 as a library to be able to keep it updated.

This patch does not revert other changes on #4206; Jaeger binaries would
continue to be built with Go 1.20 and benefit from all performance and security
improvements that the 1.20 release cycle offers.

Signed-off-by: Pablo Baeyens <pablo.baeyens@datadoghq.com>