Rebuild jaeger-ui if the tree does not match any tag exactly
The previous behavior resulted in the prebuild upstream package
being downloaded, even if the local tree has any changes,
whether they are commited or not. This broke our internal
build that has some patches applied. See:

* https://github.com/jaegertracing/jaeger/pull/4251#issuecomment-1608393236

Let's use a more strict approach and only use the prebuild package
if the tag is exact and does not contain any extra changes.

Signed-off-by: Ivan Babrou <github@ivan.computer>