ci: Keep the UI asset's .gz file timestamps the same as the original file (#4879)
## Which problem is this PR solving?
- `make rebuild-ui` and `make build-ui` did not update the UI assets
according to the latest release

## Description of the changes
- This changes the timestamp of the copied `index.html.gz` asset file to
match the timestamp of the original asset. In turn, the timestamp of
that asset matches the date of the UI release. This allows the
combination of rebuild/build targets to always bring the UI assets in
sync with the actual commit pinned in the ui-submodule.

## How was this change tested?
- validated that CI runs normally
- ran manually on different versions of UI to ensure assets get copied
again

---------

Signed-off-by: Yuri Shkuro <github@ysh.us>