Split unit-test workflow into tests and lint (#4933)
This pr splits the unit-test workflow into two parallel workflows: one
for running tests and another for running linters. The motivation behind
this change is to improve the overall workflow performance. The
unit-test workflow had become the longest among all workflows, with
significant time spent on installing tools that are only required for
linters.
Resolves #4930

---------

Signed-off-by: MeenuyD <meenu.coninja@gmail.com>
Signed-off-by: Meenu Yadav <116630390+MeenuyD@users.noreply.github.com>
Co-authored-by: Yuri Shkuro <yurishkuro@users.noreply.github.com>