Added missing tests for jaegerquery-extension
This commit includes the implementation of missing tests for jaegerquery
server defined in the `internal/extension/jaegerquery` package.

With the addition of these tests, the code coverage of this package
bumps to >95%

This commit also incorporates a small change in the jaegerstorage
extension. The `GetStorageFactory` method will now accept `StorageExt`
interface from the host. This allowed more comprehensive testing using
mock objects.

Signed-off-by: VaibhavMalik4187 <vaibhavmalik2018@gmail.com>