Make it possible to block on a fall server shutdown
The previous shutdown flow worked like:

1. Call server.Closer()
2. This would cause the http, grpc, cmux servers to stop and write
   a message to server.unavailableChannel
3. Then you would need another goroutine to drain the
   server.unavailableChannel in order for the server goroutines
   to exit

The problem is the goroutine in 3 would leak because it would
continue to block waitinf for any future writes.

This change introduces another goroutine to detect when the
server goroutines exit via a waitgroup, and then close
server.unavailableChannel once they have. This means that the
goroutine in 3. can unblock because the channel it is reading from
is clodes.

This change also ensures we are currently shutting down the server
correctly in all test code. It follows the existing pattern of
draining the server.unavailableChannel with a combination of a
wait group of 1 item and a using once.Do. I think this could be
simplified by switching to channels, but I think it would be
better to leave this for future work to be done in a single pass.

Signed-off-by: Will Sewell <willsewell@monzo.com>