Make (*Server).Stop block until servers return
This simplifies the callers in the tests because they don't need
to implement this themselves.

It also means we can remove the additional goroutine spawned
purely to close the unavailableChannel.

It does change the behaviour of (*Server).Stop to be blocking,
so it's now important that any callers _must_ consume the
unavailableChannel if they are going to call `Stop`.

Signed-off-by: Will Sewell <willsewell@monzo.com>