Refactor healthcheck signalling between server and service
Simplifies the signalling of healthcheck status from the "servers"
to the "service": instead of using 2 channels to feed healthcheck
status back to the Service.HealthCheck, we just give the server
components direct access to the Healthcheck which they can update
directly. This is possible because the Healthcheck package is
threadsafe (uses `atomic.Value` for state).

This pattern is consisten with how the service's Healtcheck is
passed directly to cmd/collector/app package.

closes #5307

Signed-off-by: Will Sewell <willsewell@monzo.com>