Fix goroutine leaks in plugin/sampling/strategystore/adaptive
This mainly involved ensuring that all goroutines started by the
Processor are shut down in a Close method (which also blocks on
them returning via a WaitGroup).

Adding this flagged an issue where the `runUpdateProbabilitiesLoop`
had a long delay, so tests need to be able to override the default
Processor.followerRefreshInterval, or they take a long time to run.

Signed-off-by: Will Sewell <willsewell@monzo.com>