Fix goroutine leaks in grpc integration tests
The key issue was that the grpc client was only being killed in
a `runtime.SetFinalizer` - i.e. when it is GCed. I think in the
tests this was not shutting down before goleak had decided that
the goroutine had leaked.

This change switches to a more conventional approach of killing
the client as part of the Close method and then ensuring this
is called in each of the tests.

While this change does not include a call to active goleak, it
was tested with this. It does not include goleak detection
because there are still other violations related to elasticsearch
in this package.

Signed-off-by: Will Sewell <willsewell@monzo.com>