[fix][query] Remove bifurcation for grpc query server (#6145)
## Which problem is this PR solving?
- Resolves #6026 

## Description of the changes
- This PR removes the bifurcation in the creation of the GRPC query
server. This bifurcation was initially added to handle the case where
users were leveraging the same port for starting both the GRPC and HTTP
query servers. However, that part is handled by `initListener` and not
at the time of creating the server, so we can simply use OTEL's helper
to create the server while leaving our logic to start the listeners
using cmux if the ports are the same.

## How was this change tested?
- CI

## Checklist
- [x] I have read
https://github.com/jaegertracing/jaeger/blob/master/CONTRIBUTING_GUIDELINES.md
- [x] I have signed all commits
- [x] I have added unit tests for the new functionality
- [x] I have run lint and test steps successfully
  - for `jaeger`: `make lint test`
  - for `jaeger-ui`: `yarn lint` and `yarn test`

Signed-off-by: Mahad Zaryab <mahadzaryab1@gmail.com>