Log time settings in CI (#6140)
## Which problem is this PR solving?
- ES v8.0/v2 integration tests often fail
([exampe](https://github.com/jaegertracing/jaeger/actions/runs/11622432732/job/32429220258)),
but looks like maybe only during certain hours of the day. E.g.
yesterday I rerun the workflow 4 times at night and it repeatedly
failed, then in the morning it succeeded.

## Description of the changes
- Log time settings in the runner to verify whether it's a time-zone
related issue with locating the indices
- Add debug logging to ES to indicate which indices are being used for
reading / writing (did not notice any issues with that in the failed run
above)
- Add debug logging to v2 binary configs
- Fix a bug in `integration.go` where `require.NoError` was using inside
lambda functions passed to `waitForCondition` - it was causing tests to
fail on transient error with Elastic. Change it to log the error and
return `false` so that `waitForCondition` may continue looping & waiting
for successful check.

---------

Signed-off-by: Yuri Shkuro <github@ysh.us>