[chore][deps] Bump otel-collector to 0.113.0 and fix failing unit test (#6197)
## Which problem is this PR solving?
- Continuation of https://github.com/jaegertracing/jaeger/pull/6175.
- There was a breaking change made in
https://github.com/open-telemetry/opentelemetry-collector/pull/11575
that updates `ToClientConn` to use `DialContext` instead of `NewClient`
because of a bug. `DialContext` uses the `passthrough` scheme was
resulting in the following error when the endpoint is left empty. This
PR fixes the unit tests by setting the endpoint in the test configs to
be non-empty.
```
grpc storage builder failed to create a store: error creating traced remote storage client: failed to exit idle mode: passthrough: received empty target in Build()
```

## How was this change tested?
- CI

## Checklist
- [x] I have read
https://github.com/jaegertracing/jaeger/blob/master/CONTRIBUTING_GUIDELINES.md
- [x] I have signed all commits
- [x] I have added unit tests for the new functionality
- [x] I have run lint and test steps successfully
  - for `jaeger`: `make lint test`
  - for `jaeger-ui`: `yarn lint` and `yarn test`

---------

Signed-off-by: dependabot[bot] <support@github.com>
Signed-off-by: Mahad Zaryab <mahadzaryab1@gmail.com>
Signed-off-by: Mahad Zaryab <43658574+mahadzaryab1@users.noreply.github.com>
Co-authored-by: dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>