Add go leak checks to ES/OS tests (#6339)
## Which problem is this PR solving?
- Fixes a part of #5006 

## Description of the changes
- This is a better version of `elasticsearch_test` reducing the number
of leaks from 50 to less than 10. Unfortunately we can't fix all the
leakages because there is no way to close the v8 client of ES leading to
some unclosed http transports. If the changes looks good then we can
merge without embedding `goleak` (as it will lead to failure of test)

## How was this change tested?
- By running the integration tests

## Checklist
- [x] I have read
https://github.com/jaegertracing/jaeger/blob/master/CONTRIBUTING_GUIDELINES.md
- [x] I have signed all commits
- [x] I have added unit tests for the new functionality
- [ ] I have run lint and test steps successfully
  - for `jaeger`: `make lint test`
  - for `jaeger-ui`: `yarn lint` and `yarn test`

---------

Signed-off-by: Manik2708 <mehtamanik96@gmail.com>