[refactor][storage] Remove metrics factory distinctions from storage factories (#6526)
## Which problem is this PR solving?
- Towards #6065

## Description of the changes
- Some of the storage factories were currently adding the `role` tag to
the metrics factory that it was getting to differentiate between
`primary` and `archive`. Given that goal in v2 is not have that
distinction, the adding of the `primary` tag was moved to the v1
meta-factory. This tag can then be overriden by the archive specific
methods in the storage factories (mainly `CreateArchiveSpanReader` and
`CreateArchiveSpanWriter`) with `role=archive`.

## How was this change tested?
- CI

## Checklist
- [x] I have read
https://github.com/jaegertracing/jaeger/blob/master/CONTRIBUTING_GUIDELINES.md
- [x] I have signed all commits
- [x] I have added unit tests for the new functionality
- [x] I have run lint and test steps successfully
  - for `jaeger`: `make lint test`
  - for `jaeger-ui`: `npm run lint` and `npm run test`

---------

Signed-off-by: Mahad Zaryab <mahadzaryab1@gmail.com>