[v2][query] Apply "Max Clock Skew Adjust" setting
Apply the `max_clock_skew_adjust` setting to the query service during
initialization.

The maximum clock skew adjustment setting for the query component is
properly read during initialization, but not applied in the v2 server in
the end. This causes it to always default to 1 second, regardless of the
configuration value.

Signed-off-by: Dominik Nakamura <dnaka91@gmail.com>