[v2][query] Apply "Max Clock Skew Adjust" setting (#6566)
Apply the `max_clock_skew_adjust` setting to the query service during
initialization.

The maximum clock skew adjustment setting for the query component is
properly read during initialization, but not applied in the v2 server in
the end. This causes it to always default to 1 second, regardless of the
configuration value.

## Which problem is this PR solving?

- Resolves #6565 

## Description of the changes

- Initialize and set the `Adjuster` on both the v1 and v2 query options.
- The config defaults to `0` so applying this unconditionally would now
disable the clock skew adjustment instead of defaulting to `1s`?
However, the same _seems to be done_ in the older services from v1.
- Having a default of `1s` instead of `0` (which disables adjustments
completely) might have been an accident?

## How was this change tested?

- Not testing so far.

## Checklist
- [x] I have read
https://github.com/jaegertracing/jaeger/blob/master/CONTRIBUTING_GUIDELINES.md
- [x] I have signed all commits
- [x] I have added unit tests for the new functionality
- [x] I have run lint and test steps successfully
  - for `jaeger`: `make lint test`
  - for `jaeger-ui`: `npm run lint` and `npm run test`

Signed-off-by: Dominik Nakamura <dnaka91@gmail.com>