Add pprof extension (#7073)
## Which problem is this PR solving?

Currently it's not possible to profile Jaeger. This makes it hard to
find bugs, memory leaks or problems in the running application. It looks
like this was somehow supported in Jaeger v1 (I found references there)
and was mentioned in `config.yaml` - but it didn't work out of the box
because the extension was missing.

## Description of the changes

This PR adds the
[pprof](https://github.com/open-telemetry/opentelemetry-collector-contrib/blob/main/extension/pprofextension/README.md)

## How was this change tested?

Using `all-in-one.yaml` and connecting to `localhost:17777`

## Checklist
- [X] I have read
https://github.com/jaegertracing/jaeger/blob/master/CONTRIBUTING_GUIDELINES.md
- [X] I have signed all commits
- [ ] I have added unit tests for the new functionality
- [X] I have run lint and test steps successfully
  - for `jaeger`: `make lint test`
  - for `jaeger-ui`: `npm run lint` and `npm run test`

---------

Signed-off-by: Denys Vitali <denys@denv.it>
Signed-off-by: amol-verma-allen <amol.verma@allen.in>