fix(elasticsearch): respect explicitly configured replicas=0 in index… (#7160)
Previously, replicas set to 0 in the OpenTelemetryCollector config
would be overwritten by default values.

This commit updates the IndexOptions struct to use a pointer type for
`Replicas` and adjusts the defaulting logic to only apply defaults when
`Replicas` is nil. This allows users to explicitly configure
`replicas: 0` without being ignored.

Reesolves #7123

---------

Signed-off-by: Masih Khatibzadeh <masih78khatibzadeh@gmail.com>