feat: add option to disable Elasticsearch health check (#7212)
## Which problem is this PR solving?
On AWS OpenSearch, I am experiencing an issue with the ElasticSearch
storage where Jaeger fails to initialize the storage on the first try.

The corresponding error message is
```
2025-06-03T13:53:19.517Z        info    jaegerstorage/extension.go:159  Initializing storage 'aws_opensearch'   {"otelcol.component.id": "jaeger_storage", "otelcol.component.kind": "extension"}
2025-06-03T13:53:25.518Z        error   extensions/extensions.go:58     Failed to start extension       {"otelcol.component.id": "jaeger_storage", "otelcol.component.kind": "extension", "error": "failed to initialize storage 'aws_opensearch': failed to create Elasticsearch client: Head \"https://my-opensearch-domain.es.amazonaws.com:443\": context deadline exceeded"}
```

## Description of the changes
The problem was already reported in the olivere/elastic library that
Jaeger uses: https://github.com/olivere/elastic/issues/880. The proposed
fix is to disable the health check. Currently, this is possible in
Jaeger by setting `auth.bearer_token.from_context = true` as a
workaround.
The change in this PR adds a configuration option,
`disable_health_check`, that can be used to disable the health check
directly.

## How was this change tested?
I tested the change locally with Jaeger v1 and v2, logging the value
that is used in `elastic.SetHealthcheck`. I tested the workaround
mentioned above in our test environment, where it solved the problem.

## Checklist
- [x] I have read
https://github.com/jaegertracing/jaeger/blob/master/CONTRIBUTING_GUIDELINES.md
- [x] I have signed all commits
- [x] I have added unit tests for the new functionality
- [x] I have run lint and test steps successfully
  - for `jaeger`: `make lint test`
  - for `jaeger-ui`: `npm run lint` and `npm run test`

Signed-off-by: Timon Engelke <timon.engelke@inovex.de>